Some refactoring based on pr comments
