+++ samza-core/src/main/java/org/apache/samza/execution/StreamEdge.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.samza.processorgraph;
+package org.apache.samza.execution;
 
 import java.util.ArrayList;
 import java.util.List;
