+++ samza-core/src/test/java/org/apache/samza/processorgraph/TestProcessorGraph.java
@@ -23,6 +23,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import org.apache.samza.execution.ProcessorGraph;
+import org.apache.samza.execution.ProcessorNode;
 import org.apache.samza.system.StreamSpec;
 import org.junit.Before;
 import org.junit.Test;
